Stats: Refactoring controller and page files
