+++ client/my-sites/stats/controller.jsx
@@ -1,6 +1,5 @@
 import config from '@automattic/calypso-config';
 import page from '@automattic/calypso-router';
-import { Spinner } from '@automattic/components';
 import i18n from 'i18n-calypso';
 import { find, pick } from 'lodash';
 import moment from 'moment';
@@ -11,53 +10,14 @@ import { getSite, getSiteOption } from 'calypso/state/sites/selectors';
 import { setNextLayoutFocus } from 'calypso/state/ui/layout-focus/actions';
 import { getCurrentLayoutFocus } from 'calypso/state/ui/layout-focus/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import { rangeOfPeriod, getSiteFilters } from './pages/shared/helpers';
+import PageLoading from './pages/shared/page-loading';
 import StatsSite from './site';
 import StatsEmailDetail from './stats-email-detail';
 import StatsEmailSummary from './stats-email-summary';
 import LoadStatsPage from './stats-redirect/load-stats-page';
 
-const isOdysseyStats = config.isEnabled( 'is_running_in_jetpack_site' );
 const isPurchaseFlowEnabled = config.isEnabled( 'stats/checkout-flows-v2' );
-
-const PageLoading = (
-	<div
-		style={ {
-			minHeight: 'calc( 100vh - 100px )',
-			width: '100%',
-			display: 'flex',
-			justifyContent: 'center',
-			alignItems: 'center',
-		} }
-	>
-		{ isOdysseyStats ? (
-			<img width="32" height="32" alt="Loading" src="//en.wordpress.com/i/loading/loading-64.gif" />
-		) : (
-			<Spinner />
-		) }
-	</div>
-);
-
-function rangeOfPeriod( period, date ) {
-	const periodRange = {
-		period: period,
-		startOf: date.clone().startOf( period ),
-		endOf: date.clone().endOf( period ),
-	};
-
-	if ( 'week' === period ) {
-		if ( '0' === date.format( 'd' ) ) {
-			periodRange.startOf.subtract( 6, 'd' );
-			periodRange.endOf.subtract( 6, 'd' );
-		} else {
-			periodRange.startOf.add( 1, 'd' );
-			periodRange.endOf.add( 1, 'd' );
-		}
-	}
-
-	periodRange.key = period + ':' + periodRange.endOf.format( 'YYYY-MM-DD' );
-
-	return periodRange;
-}
 
 function getNumPeriodAgo( momentSiteZone, date, period ) {
 	const endOfCurrentPeriod = momentSiteZone.endOf( period );
@@ -79,108 +39,6 @@ function getNumPeriodAgo( momentSiteZone, date, period ) {
 			break;
 	}
 	return numPeriodAgo;
-}
-
-function getSiteFilters( siteId ) {
-	return [
-		//TODO: This Insights route could be removed since it has been set routing as below.
-		// statsPage( '/stats/insights/:site', insights );
-		{
-			title: i18n.translate( 'Insights' ),
-			path: '/stats/insights/' + siteId,
-			id: 'stats-insights',
-		},
-		{
-			title: i18n.translate( 'Subscribers' ),
-			path: '/stats/subscribers/' + siteId,
-			id: 'stats-subscribers',
-			period: 'day', // default period for Subscribers
-		},
-		{
-			title: i18n.translate( 'Days' ),
-			path: '/stats/subscribers/day/' + siteId,
-			id: 'stats-subscribers-day',
-			period: 'day',
-		},
-		{
-			title: i18n.translate( 'Weeks' ),
-			path: '/stats/subscribers/week/' + siteId,
-			id: 'stats-subscribers-week',
-			period: 'week',
-		},
-		{
-			title: i18n.translate( 'Months' ),
-			path: '/stats/subscribers/month/' + siteId,
-			id: 'stats-subscribers-month',
-			period: 'month',
-		},
-		{
-			title: i18n.translate( 'Years' ),
-			path: '/stats/subscribers/year/' + siteId,
-			id: 'stats-subscribers-year',
-			period: 'year',
-		},
-		{
-			title: i18n.translate( 'Hours' ),
-			path: '/stats/hour/' + siteId,
-			id: 'stats-hour',
-			period: 'hour',
-		},
-		{
-			title: i18n.translate( 'Days' ),
-			path: '/stats/day/' + siteId,
-			id: 'stats-day',
-			period: 'day',
-		},
-		{
-			title: i18n.translate( 'Weeks' ),
-			path: '/stats/week/' + siteId,
-			id: 'stats-week',
-			period: 'week',
-		},
-		{
-			title: i18n.translate( 'Months' ),
-			path: '/stats/month/' + siteId,
-			id: 'stats-month',
-			period: 'month',
-		},
-		{
-			title: i18n.translate( 'Years' ),
-			path: '/stats/year/' + siteId,
-			id: 'stats-year',
-			period: 'year',
-		},
-		{
-			title: i18n.translate( 'Hours' ),
-			path: `/stats/email/opens/hour/`,
-			id: 'stats-email-opens-hour',
-			period: 'hour',
-		},
-		{
-			title: i18n.translate( 'Days' ),
-			path: `/stats/email/opens/day/`,
-			id: 'stats-email-opens-day',
-			period: 'day',
-		},
-		{
-			title: i18n.translate( 'Hours' ),
-			path: `/stats/email/clicks/hour/`,
-			id: 'stats-email-clicks-hour',
-			period: 'hour',
-		},
-		{
-			title: i18n.translate( 'Days' ),
-			path: `/stats/email/clicks/day/`,
-			id: 'stats-email-clicks-day',
-			period: 'day',
-		},
-		{
-			title: i18n.translate( 'Days' ),
-			path: `/stats/day/emails`,
-			id: 'stats-email-summary',
-			period: 'day',
-		},
-	];
 }
 
 function getWordAdsFilters( siteId ) {
@@ -247,48 +105,6 @@ export function redirectToDefaultWordAdsPeriod( context ) {
 
 export function redirectToDefaultModulePage( context ) {
 	page.redirect( `/stats/day/${ context.params.module }/${ context.params.site }` );
-}
-
-export function insights( context, next ) {
-	context.primary = isPurchaseFlowEnabled ? (
-		<LoadStatsPage>
-			<AsyncLoad require="calypso/my-sites/stats/stats-insights" placeholder={ PageLoading } />
-		</LoadStatsPage>
-	) : (
-		<AsyncLoad require="calypso/my-sites/stats/stats-insights" placeholder={ PageLoading } />
-	);
-	next();
-}
-
-export function subscribers( context, next ) {
-	const givenSiteId = context.params.site;
-	const filters = getSiteFilters( givenSiteId );
-	const activeFilter = find( filters, ( filter ) => {
-		return (
-			context.path.indexOf( filter.path ) >= 0 ||
-			( filter.altPaths && context.path.indexOf( filter.altPaths ) >= 0 )
-		);
-	} );
-
-	// moment and rangeOfPeriod format needed for summary page link for email mdule
-	const date = moment().locale( 'en' );
-
-	context.primary = isPurchaseFlowEnabled ? (
-		<LoadStatsPage>
-			<AsyncLoad
-				require="calypso/my-sites/stats/stats-subscribers"
-				placeholder={ PageLoading }
-				period={ rangeOfPeriod( activeFilter.period, date ) }
-			/>
-		</LoadStatsPage>
-	) : (
-		<AsyncLoad
-			require="calypso/my-sites/stats/stats-subscribers"
-			placeholder={ PageLoading }
-			period={ rangeOfPeriod( activeFilter.period, date ) }
-		/>
-	);
-	next();
 }
 
 export function overview( context, next ) {
@@ -744,14 +560,6 @@ export function emailSummary( context, next ) {
 	next();
 }
 
-export function purchase( context, next ) {
-	context.primary = (
-		// DO NOT WRAP WITH <LoadStatsPage /> or you will get an infinite loop
-		<AsyncLoad
-			require="calypso/my-sites/stats/stats-purchase"
-			placeholder={ PageLoading }
-			query={ context.query }
-		/>
-	);
-	next();
-}
+export { default as insights } from './pages/insights/controller';
+export { default as subscribers } from './pages/subscribers/controller';
+export { default as purchase } from './pages/purchase/controller';
